server/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
certifi==2024.12.14
click==8.1.8
colorama==0.4.6
Deprecated==1.2.15
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
limits==3.14.1
Mako==1.3.8
MarkupSafe==3.0.2
openai==1.58.1
packaging==24.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
wrapt==1.17.0
